food-detection-api/requirements.txt
flask==3.0.0
ultralytics==8.4.36
pillow==10.3.0
gunicorn==21.2.0
